backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# MCP - require a newer mcp (with cli extras) to satisfy mcp-tools / mcp-server
mcp[cli]>=1.12.3,<2.0

# File Processing
python-docx==1.1.2
markdown==3.7
chardet==5.2.0
watchdog==6.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx>=0.28.1
